fix- tentativa de correção de bugs da service de requisições

+++ src/main/kotlin/school/sptech/projetotfg/service/Caso.kt
@@ -2,7 +2,7 @@ package school.sptech.projetotfg.service
 
 data class Caso(
     val assunto:String,
-    val inicioTri:Int,
-    val fimTri:Int,
-    val situacao:Int
+    val inicioTri:Long,
+    val fimTri:Long,
+    val situacao:Long
 )
